linh tinh

+++ rooms/src/main/java/com/thuc/rooms/repository/TripRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TripRepository extends JpaRepository<Trip, Long> {
+public interface TripRepository extends JpaRepository<Trip, Integer> {
 
     List<Trip> findByTripTypeContainingIgnoreCase(String trip);
 }
